# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AppliedFilters": {
        "description": "The filters actually applied after validation and defaults (default window: 30 days\nending at the database's latest trending date). Empty country/category lists mean all.",
        "properties": {
          "categories": {
            "description": "Applied categories; empty = all categories.",
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "countries": {
            "description": "Applied country codes; empty = all countries.",
            "items": {
              "type": "string"
            },
            "title": "Countries",
            "type": "array"
          },
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "end_date": {
            "format": "date",
            "title": "End Date",
            "type": "string"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          }
        },
        "required": [
          "start_date",
          "end_date",
          "days",
          "countries",
          "categories"
        ],
        "title": "AppliedFilters",
        "type": "object"
      },
...
      "CurrentUser": {
        "properties": {
          "tenant": {
            "$ref": "#/components/schemas/TenantOut"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "required": [
          "user",
          "tenant"
        ],
        "title": "CurrentUser",
        "type": "object"
      },
...
      "DailySplit": {
        "enum": [
          "country",
          "category"
        ],
        "title": "DailySplit",
        "type": "string"
      },
      "DailyVolume": {
        "description": "Daily trending activity. No daily views: view counts are cumulative lifetime totals.",
        "properties": {
          "series": {
            "items": {
              "$ref": "#/components/schemas/DailySeries"
            },
            "title": "Series",
            "type": "array"
          },
          "split_by": {
            "anyOf": [
              {
                "enum": [
                  "country",
                  "category"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Split By"
          }
        },
        "required": [
          "split_by",
          "series"
        ],
        "title": "DailyVolume",
        "type": "object"
      },
      "DataResponse_CurrentUser_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CurrentUser"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[CurrentUser]",
        "type": "object"
      },
...
      "DataResponse_LogoutResult_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LogoutResult"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[LogoutResult]",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Same value as the X-Request-ID response header.",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FilteredResponse_DailyVolume_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DailyVolume"
          },
          "filters": {
            "$ref": "#/components/schemas/AppliedFilters"
          }
        },
        "required": [
          "filters",
          "data"
        ],
        "title": "FilteredResponse[DailyVolume]",
        "type": "object"
      },
...
      "FilteredResponse_Kpis_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Kpis"
          },
          "filters": {
            "$ref": "#/components/schemas/AppliedFilters"
          }
        },
        "required": [
          "filters",
          "data"
        ],
        "title": "FilteredResponse[Kpis]",
        "type": "object"
      },
...
      "Kpis": {
        "description": "Overview KPIs for the filtered period and the immediately preceding period of equal length\n(same countries and categories).",
        "properties": {
          "engagement_rate": {
            "$ref": "#/components/schemas/RateMetric",
            "description": "VIDEO-LEVEL: (SUM(likes) + SUM(comments)) / SUM(views) over the same latest snapshots. Change is in percentage points."
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "previous_period": {
            "$ref": "#/components/schemas/PreviousPeriod"
          },
          "trending_volume": {
            "$ref": "#/components/schemas/CountMetric",
            "description": "SNAPSHOT-LEVEL: number of filtered snapshot rows (country x video x day appearances)."
          },
          "unique_channels": {
            "$ref": "#/components/schemas/CountMetric",
            "description": "SNAPSHOT-LEVEL: distinct channels among filtered snapshots."
          },
          "unique_videos": {
            "$ref": "#/components/schemas/CountMetric",
            "description": "VIDEO-LEVEL: COUNT(DISTINCT video_id) among filtered snapshots."
          },
          "views": {
            "$ref": "#/components/schemas/CountMetric",
            "description": "VIDEO-LEVEL: sum of view_count from each video's latest snapshot within the filtered dataset, counted once per video. Views are never summed across historical snapshots. View counts are lifetime totals, so the change compares the videos trending in each period, not views gained."
          }
        },
        "required": [
          "period",
          "previous_period",
          "unique_videos",
          "trending_volume",
          "views",
          "unique_channels",
          "engagement_rate"
        ],
        "title": "Kpis",
        "type": "object"
      },
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "maxLength": 320,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LogoutResult": {
        "properties": {
          "status": {
            "const": "logged_out",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "LogoutResult",
        "type": "object"
      },
...
      "Period": {
        "properties": {
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "end_date": {
            "format": "date",
            "title": "End Date",
            "type": "string"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          }
        },
        "required": [
          "start_date",
          "end_date",
          "days"
        ],
        "title": "Period",
        "type": "object"
      },
...
      "RegisterRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "maxLength": 320,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "description": "12 to 128 characters.",
            "maxLength": 1024,
            "title": "Password",
            "type": "string"
          },
          "tenant_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Workspace name. Default: '<display_name>'s workspace'.",
            "title": "Tenant Name"
          }
        },
        "required": [
          "email",
          "password",
          "display_name"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "Role": {
        "enum": [
          "OWNER",
          "ADMIN",
          "MEMBER"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "TenantOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "TenantOut",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "permissions": {
            "description": "Derived from the role on the server.",
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "title": "Permissions",
            "type": "array"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "role",
          "permissions"
        ],
        "title": "UserOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nRead-only analytics API over YouTube trending snapshots (PostgreSQL `app` schema) and a frozen ML model.\n\n**Metric semantics**\n* *Latest snapshot* = each video's latest snapshot **within the filtered dataset** (date range,\n  countries, categories), not its globally latest snapshot.\n* **Snapshot-level** metrics count trending appearances (country x video x day), e.g. trending volume.\n* **Video-level** metrics count each video once via its latest filtered snapshot, e.g. unique videos,\n  views, engagement rate.\n* **Country-level** metrics use each video's latest filtered snapshot in that country. Country\n  `views_of_trending_videos` are **global lifetime views** of videos that trended there; they are\n  **not additive** across countries.\n* View counts are lifetime totals: historical snapshots are **never summed** for views.\n* Period change: counts and views in **percent**; engagement rate in **percentage points**.\n\n**Authentication**: sign in via `/api/v1/auth/login` (or register). The session is an HttpOnly\ncookie; every `/api/v1` endpoint except register/login requires it. Analytics endpoints need\n`VIEW_ANALYTICS`, predictions `MAKE_PREDICTION`, member management `MANAGE_USERS`. The YouTube\ndataset is global and read-only; workspaces (tenants) own only their users and memberships.\nHealth probes are public.\n",
    "title": "YouTube Trending Intelligence API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_CurrentUser_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid email or password."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Account temporarily locked."
          }
        },
        "summary": "Sign in with email and password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_LogoutResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          }
        },
        "summary": "Sign out: revoke the current session and clear the cookie (idempotent)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_CurrentUser_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in or session expired."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          }
        },
        "summary": "The signed-in user, role and workspace",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_CurrentUser_"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Registration is disabled."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Email already registered."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          }
        },
        "summary": "Create an account and a new workspace (you become its OWNER) and sign in",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/overview/daily-volume": {
      "get": {
        "operationId": "daily_volume_api_v1_overview_daily_volume_get",
        "parameters": [
          {
            "description": "Split into one series per country or per (snapshot) category.",
            "in": "query",
            "name": "split_by",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DailySplit"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Split into one series per country or per (snapshot) category.",
              "title": "Split By"
            }
          },
          {
            "description": "First trending date (inclusive, YYYY-MM-DD). Default: 29 days before end_date, not earlier than the first date in the data.",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First trending date (inclusive, YYYY-MM-DD). Default: 29 days before end_date, not earlier than the first date in the data.",
              "title": "Start Date"
            }
          },
          {
            "description": "Last trending date (inclusive). Default: the latest trending date in the data.",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last trending date (inclusive). Default: the latest trending date in the data.",
              "title": "End Date"
            }
          },
          {
            "description": "Country code; repeat for several (?country=IN&country=US). Omit for all countries.",
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maxLength": 64,
                    "type": "string"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Country code; repeat for several (?country=IN&country=US). Omit for all countries.",
              "title": "Country"
            }
          },
          {
            "description": "Category (case-insensitive); repeat for several. Omit for all categories.",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maxLength": 64,
                    "type": "string"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Category (case-insensitive); repeat for several. Omit for all categories.",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilteredResponse_DailyVolume_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in or session expired."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Your role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request, filter or parameter."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server or database error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database unavailable or busy."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database statement timeout."
          }
        },
        "summary": "Daily trending volume and unique videos (zero-filled)",
        "tags": [
          "overview"
        ]
      }
    },
    "/api/v1/overview/kpis": {
      "get": {
        "description": "'Latest snapshot' means each video's latest snapshot WITHIN THE FILTERED DATASET (the rows matching the date range, countries and categories), not its globally latest snapshot. Counts and views change in percent; engagement_rate changes in percentage points. The previous period is the immediately preceding period of equal length with the same countries and categories.",
        "operationId": "kpis_api_v1_overview_kpis_get",
        "parameters": [
          {
            "description": "First trending date (inclusive, YYYY-MM-DD). Default: 29 days before end_date, not earlier than the first date in the data.",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First trending date (inclusive, YYYY-MM-DD). Default: 29 days before end_date, not earlier than the first date in the data.",
              "title": "Start Date"
            }
          },
          {
            "description": "Last trending date (inclusive). Default: the latest trending date in the data.",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last trending date (inclusive). Default: the latest trending date in the data.",
              "title": "End Date"
            }
          },
          {
            "description": "Country code; repeat for several (?country=IN&country=US). Omit for all countries.",
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maxLength": 64,
                    "type": "string"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Country code; repeat for several (?country=IN&country=US). Omit for all countries.",
              "title": "Country"
            }
          },
          {
            "description": "Category (case-insensitive); repeat for several. Omit for all categories.",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maxLength": 64,
                    "type": "string"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Category (case-insensitive); repeat for several. Omit for all categories.",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilteredResponse_Kpis_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in or session expired."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Your role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request, filter or parameter."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server or database error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database unavailable or busy."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database statement timeout."
          }
        },
        "summary": "Overview KPIs with period-over-period change",
        "tags": [
          "overview"
        ]
      }
    },
...
  }
}